tradesage-backend/market_data/requirements.txt
databento>=0.30.0
asyncpg>=0.29.0
redis>=5.0.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0